# docs/openapi.yml
openapi: 3.0.3
info:
  title: Stellar GreenPay API
  version: 1.0.0
  description: |
    REST API for the Stellar GreenPay climate donation platform.
    Every donation is recorded on the Stellar blockchain via Soroban smart contracts.
    All `/api/v1/*` JSON responses use a single envelope. Successful responses
    are `{ success: true, data, meta? }`; error responses are
    `{ success: false, error: { code, message, details? } }`. The unversioned
    `/api/*` paths are legacy aliases that redirect to `/api/v1/*`; clients
    should migrate to `/api/v1/*` and branch on `error.code`, never message text.
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development
...
components:
  schemas:
    Success:
      type: object
      required: [success, data]
      additionalProperties: false
      properties:
        success:
          type: boolean
          enum: [true]
          example: true
        data:
          description: Endpoint-specific response payload.
        meta:
          type: object
          additionalProperties: true
          description: Optional response metadata such as pagination or cursors.
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [Reforestation, Solar Energy, Ocean Conservation, Clean Water, Wildlife Protection, Carbon Capture, Wind Energy, Sustainable Agriculture, Other]
        status:
          type: string
          enum: [active, completed, paused, rejected]
        location:
          type: string
        walletAddress:
          type: string
        raisedXLM:
          type: string
        donorCount:
          type: integer
        co2OffsetKg:
          type: number
        verified:
          type: boolean
        onChainVerified:
          type: boolean
        tags:
          type: array
          items:
            type: string
        aiSummary:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProjectListMeta:
      allOf:
        - $ref: '#/components/schemas/ProjectSearchMeta'
...
paths:
...
  /api/projects:
    get:
      tags: [Projects]
      summary: List projects
      operationId: listProjects
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [Reforestation, Solar Energy, Ocean Conservation, Clean Water, Wildlife Protection, Carbon Capture, Wind Energy, Sustainable Agriculture, Other]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, paused]
        - name: verified
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of projects with optional search metadata
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Project'
                      meta:
                        $ref: '#/components/schemas/ProjectListMeta'
  /api/projects/featured:
    get:
      tags: [Projects]
      summary: Get featured project
      operationId: getFeaturedProject
      description: Returns the active project with the highest donor count. Cached for 24 hours.
      responses:
        '200':
          description: Featured project
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Project'
        '404':
          description: No featured project found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/projects/admin/register:
    post:
      tags: [Projects]
      summary: Build on-chain project registration XDR
      operationId: adminRegisterProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, name, wallet, co2PerXLM, adminAddress]
              properties:
                projectId:
                  type: string
                name:
                  type: string
                wallet:
                  type: string
                co2PerXLM:
                  type: integer
                adminAddress:
                  type: string
      responses:
        '200':
          description: XDR for admin to sign
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          xdr:
                            type: string
        '500':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/projects/admin/confirm:
    post:
      tags: [Projects]
      summary: Confirm on-chain project registration
      operationId: adminConfirmProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transactionHash, projectId]
              properties:
                transactionHash:
                  type: string
                projectId:
                  type: string
      responses:
        '200':
          description: Project updated as verified
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Project'
  /api/projects/{id}:
    get:
      tags: [Projects]
      summary: Get project by ID
      operationId: getProject
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: Project detail with campaigns, milestones, ratings
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/Project'
                          - type: object
                            properties:
                              campaigns:
                                type: array
                                items:
                                  $ref: '#/components/schemas/Campaign'
                              activeCampaign:
                                $ref: '#/components/schemas/Campaign'
                                nullable: true
                              averageRating:
                                type: number
                              ratingCount:
                                type: integer
                              milestones:
                                type: array
                                items:
                                  $ref: '#/components/schemas/Milestone'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/projects/{id}/status:
    patch:
      tags: [Projects]
      summary: Update project status
      operationId: updateProjectStatus
      parameters:
        - $ref: '#/components/parameters/projectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [active, rejected, paused]
                reason:
                  type: string
                adminAddress:
                  type: string
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Project'
        '400':
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/projects/{id}/verify:
    get:
      tags: [Projects]
      summary: Verify project on-chain
      operationId: verifyProject
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: On-chain verification result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          projectId:
                            type: string
                          onChainVerified:
                            type: boolean
                          contractRegisteredAt:
                            type: integer
                            nullable: true
                          totalRaisedOnChain:
                            type: string
  /api/projects/{id}/campaigns:
    get:
      tags: [Projects]
      summary: List campaigns for a project
      operationId: listProjectCampaigns
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: List of campaigns
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Campaign'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Projects]
      summary: Create a campaign for a project
      operationId: createProjectCampaign
      parameters:
        - $ref: '#/components/parameters/projectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, goalXLM, deadline]
              properties:
                title:
                  type: string
                  minLength: 3
                  maxLength: 120
                goalXLM:
                  type: number
                  example: 1000.0
                deadline:
                  type: string
                  format: date-time
                description:
                  type: string
                  maxLength: 500
                adminAddress:
                  type: string
      responses:
        '201':
          description: Campaign created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Campaign'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...